+++ script/core/signature.lua
@@ -114,7 +114,13 @@ local function isEventNotMatch(call, src)
     end
     local eventLiteral = event.extends.types[1] and guide.getLiteral(event.extends.types[1])
     if eventLiteral == nil then
-        return false
+        -- extra checking when function param is not pure literal
+        -- eg: it maybe an alias type with literal values
+        local eventMap = vm.getLiterals(event.extends.types[1])
+        if not eventMap then
+            return false
+        end
+        return not eventMap[literal]
     end
     return eventLiteral ~= literal
 end